Break out inlined logic
